Changed all Floats to Doubles in PepXml parser

+++ MSFileToolbox/src/umich/ms/fileio/filetypes/pepxml/jaxb/primitive/AminoacidModification.java
@@ -33,8 +33,8 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;attribute name="aminoacid" use="required" type="{http://www.w3.org/2001/XMLSchema}string" />
- *       &lt;attribute name="massdiff" use="required" type="{http://www.w3.org/2001/XMLSchema}float" />
- *       &lt;attribute name="mass" use="required" type="{http://www.w3.org/2001/XMLSchema}float" />
+ *       &lt;attribute name="massdiff" use="required" type="{http://www.w3.org/2001/XMLSchema}double" />
+ *       &lt;attribute name="mass" use="required" type="{http://www.w3.org/2001/XMLSchema}double" />
  *       &lt;attribute name="variable" use="required" type="{http://www.w3.org/2001/XMLSchema}string" />
  *       &lt;attribute name="peptide_terminus" type="{http://www.w3.org/2001/XMLSchema}string" />
  *       &lt;attribute name="protein_terminus" type="{http://www.w3.org/2001/XMLSchema}string" />
@@ -55,9 +55,9 @@ public class AminoacidModification {
     @XmlAttribute(name = "aminoacid", required = true)
     protected String aminoacid;
     @XmlAttribute(name = "massdiff", required = true)
-    protected float massdiff;
+    protected double massdiff;
     @XmlAttribute(name = "mass", required = true)
-    protected float mass;
+    protected double mass;
     @XmlAttribute(name = "variable", required = true)
     protected String variable;
     @XmlAttribute(name = "peptide_terminus")
@@ -99,7 +99,7 @@ public class AminoacidModification {
      * Gets the value of the massdiff property.
      * 
      */
-    public float getMassdiff() {
+    public double getMassdiff() {
         return massdiff;
     }
 
@@ -107,7 +107,7 @@ public class AminoacidModification {
      * Sets the value of the massdiff property.
      * 
      */
-    public void setMassdiff(float value) {
+    public void setMassdiff(double value) {
         this.massdiff = value;
     }
 
@@ -115,7 +115,7 @@ public class AminoacidModification {
      * Gets the value of the mass property.
      * 
      */
-    public float getMass() {
+    public double getMass() {
         return mass;
     }
 
@@ -123,7 +123,7 @@ public class AminoacidModification {
      * Sets the value of the mass property.
      * 
      */
-    public void setMass(float value) {
+    public void setMass(double value) {
         this.mass = value;
     }
