+++ MSFileToolbox/src/umich/ms/fileio/filetypes/pepxml/jaxb/primitive/FragmentMasses.java
@@ -33,8 +33,8 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;attribute name="channel" use="required" type="{http://regis-web.systemsbiology.net/pepXML}positiveInt" />
- *       &lt;attribute name="mz" use="required" type="{http://www.w3.org/2001/XMLSchema}float" />
- *       &lt;attribute name="offset" type="{http://www.w3.org/2001/XMLSchema}float" />
+ *       &lt;attribute name="mz" use="required" type="{http://www.w3.org/2001/XMLSchema}double" />
+ *       &lt;attribute name="offset" type="{http://www.w3.org/2001/XMLSchema}double" />
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
@@ -49,9 +49,9 @@ public class FragmentMasses {
     @XmlAttribute(name = "channel", required = true)
     protected long channel;
     @XmlAttribute(name = "mz", required = true)
-    protected float mz;
+    protected double mz;
     @XmlAttribute(name = "offset")
-    protected float offset;
+    protected double offset;
 
     /**
      * Gets the value of the channel property.
@@ -73,7 +73,7 @@ public class FragmentMasses {
      * Gets the value of the mz property.
      * 
      */
-    public float getMz() {
+    public double getMz() {
         return mz;
     }
 
@@ -81,7 +81,7 @@ public class FragmentMasses {
      * Sets the value of the mz property.
      * 
      */
-    public void setMz(float value) {
+    public void setMz(double value) {
         this.mz = value;
     }
 
@@ -89,7 +89,7 @@ public class FragmentMasses {
      * Gets the value of the offset property.
      * 
      */
-    public float getOffset() {
+    public double getOffset() {
         return offset;
     }
 
@@ -97,7 +97,7 @@ public class FragmentMasses {
      * Sets the value of the offset property.
      * 
      */
-    public void setOffset(float value) {
+    public void setOffset(double value) {
         this.offset = value;
     }
